Lay groundwork for megafauna, and add Ashdragon.
Includes a refactor for gun turrets (merged into portable turrets), hostile simple_animal behaviour, spells (sounds have been added) and poi_list items.

+++ code/modules/power/supermatter/supermatter.dm
@@ -64,6 +64,7 @@
 
 /obj/machinery/power/supermatter_shard/New()
 	. = ..()
+	poi_list |= src
 	radio = new(src)
 	radio.listening = 0
 	investigate_log("has been created.", "supermatter")
@@ -71,8 +72,10 @@
 
 /obj/machinery/power/supermatter_shard/Destroy()
 	investigate_log("has been destroyed.", "supermatter")
-	qdel(radio)
-	radio = null
+	if(radio)
+		qdel(radio)
+		radio = null
+	poi_list.Remove(src)
 	return ..()
 
 /obj/machinery/power/supermatter_shard/proc/explode()
